"Lead@ section adjusted a bit, added animations, redesigned spacing and things like that
