+++ src/components/IntroPicture/IntroPicture.module.scss
@@ -12,8 +12,8 @@
 .sketch {
     position: absolute;
     transform: translateX(5%) rotate(180deg);
-    bottom: -3.5vw;
-    width: 40vw;
+    bottom: -2vw;
+    width: 35vw;
     z-index: -2;
     transition: all 2s ease !important;
     opacity: .5;
