+++ src/components/ProjectCard/styles.module.scss
@@ -1,0 +1,12 @@
+@import '../../styles/variables';
+
+.box {
+    box-shadow: none;
+    transition: all 1s ease;
+    border-radius: 25px;
+    
+    &:hover {
+        box-shadow: 0 0.5em 1em -0.125em $grey-30p, 0 0px 0 1px $grey-10p;
+        transform: translate(0%, -5%);
+    }
+}
