+++ src/components/index.js
@@ -2,6 +2,6 @@ export {ContactPoints} from './ContactPoints/ContactPoints';
 export {Divider} from './Divider/Divider';
 export {IntroPicture} from './IntroPicture/IntroPicture';
 export {Layout} from './Layout/Layout';
-export {ProjectCard} from './ProjectCard/ProjectCard';
+export {ProjectCard} from './ProjectCard';
 export {SectionSentence} from './SectionSentence/SectionSentence';
 export {TechStackElement} from './TechStackElement/TechStackElement';
