+++ src/sections/LeadSection/LeadSection.jsx
@@ -16,15 +16,15 @@ const LeadSection = () => {
                 />
             </Columns.Column>
             <Columns.Column size={6}>
-                <h3 className="title is-size-4 mb-0 is-title-reveal">
+                <h3 className="title is-size-3 mb-0 is-title-reveal">
                     <span className='has-text-danger'>{`${new Date().getFullYear()-2008}`} years</span> in leadership
                 </h3>
-                <Content size="small">
+                <Content size="">
                     <ul>
-                        <li>Leading teams 4 to 250 <span className='has-text-weight-bold'>since 2008</span></li>
-                        <li><span className='has-text-weight-bold'>Founder</span> of multiple non-profits (9y), SMBs (2y) and Startups (7y)</li>
-                        <li>Deep motivation and practical experience in building <span className='has-text-weight-bold'>diverse, multi-language, multi-culture teams</span> (8y)</li>
-                        <li>Serial entrepreneur from heart, founded & validated <span className='has-text-weight-bold'>more than 30 projects</span>.</li>
+                        <li>Leading teams of 4 to 250 <span className='has-text-weight-bold'>since 2008</span></li>
+                        <li><span className='has-text-weight-bold'>Founder</span> of multiple non-profits (9y), SMBs (3y) and Startups (7y)</li>
+                        <li>Practical experience in building <span className='has-text-weight-bold'>diverse, multi-language, multi-culture teams</span> (8y)</li>
+                        <li>Serial entrepreneur from heart, founded and validated <span className='has-text-weight-bold'>more than 30 projects</span></li>
                     </ul>
                 </Content>
             </Columns.Column>
